@@ -10,13 +10,69 @@ namespace Analogy
         public static IEnumerable<AnalogyChangeLog> GetChangeLog()
         {
             return new List<AnalogyChangeLog>
-            {
-                new AnalogyChangeLog("V4.3.3 - [gRPC] Add self hosting gRPC server #735",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2020,12,07)),
-                new AnalogyChangeLog("V4.3.3 - [Debug] Reflect internal log changes in real time #736",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2020,12,06)),
-                new AnalogyChangeLog("V4.3.3 - Remove Debugger.Launch from open file association path #691",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2020,11,28)),
-                new AnalogyChangeLog("V4.3.3 - Move User Statistics outside of user settings #689",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2020,11,28)),
-                new AnalogyChangeLog("V4.3.3 - Auto restart real time provider selection is broken #688",AnalogChangeLogType.Bug,"Lior Banai",new DateTime(2020,11,28)),
-                new AnalogyChangeLog("V4.3.3 - [Feature] Restore Extensions code (User control) #687",AnalogChangeLogType.Feature,"Lior Banai",new DateTime(2020,11,28)),
+            {       
+                new AnalogyChangeLog("V4.7.3 - [UI] Typing non numerical text in numerical fields causes exception #823",AnalogChangeLogType.Bug,"Lior Banai",new DateTime(2021,06,22)),
+                new AnalogyChangeLog("V4.7.3 - FileNotFound: System.Drawing.Common #1023",AnalogChangeLogType.Bug,"Lior Banai",new DateTime(2021,06,22)),
+                new AnalogyChangeLog("V4.7.2 - [DevExpress] Upgrade version to V21.1 #1014",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,06,11)),
+                new AnalogyChangeLog("V4.7.2 - Local time for UTC timestamps #1016",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,06,11)),
+                new AnalogyChangeLog("V4.7.2 - Allow copy values from the json formatted view #1018",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,06,11)),
+                new AnalogyChangeLog("V4.7.2 - Watch or tail? #1015",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,06,11)),
+                new AnalogyChangeLog("V4.7.1 - [Update mechanism] Add NET5 download flag #1003",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,05,21)),
+                new AnalogyChangeLog("V4.7.1 - Add Json parser to the detailed message window #1002",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,05,21)),
+                new AnalogyChangeLog("V4.7.0 - Update version to 4.7.0 #978",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,05,01)),
+                new AnalogyChangeLog("V4.7.0 - [DevExpress 20.2] set default skin to be the same as old skin #971",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,05,01)),
+                new AnalogyChangeLog("V4.7.0 - [Net framework 4.X] Remove WCF data provider #976",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,05,01)),
+                new AnalogyChangeLog("V4.7.0 - [Analogy Updater] add partial mode for updating #977",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,05,01)),
+                new AnalogyChangeLog("V4.7.0 - [DevExpress V20.2] FluentDesignForm ignores the mouse click in Windows 10 mode #973",AnalogChangeLogType.Bug,"Lior Banai",new DateTime(2021,04,29)),
+                new AnalogyChangeLog("V4.7.0 - [NET 5] Add Net 5 target compilation #974",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,04,27)),
+                new AnalogyChangeLog("V4.7.0 - [DevExpress] Upgrade version to v20.2.6 #967",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,04,13)),
+                new AnalogyChangeLog("V4.7.0 - [chart] add zoom and scroll functionality #969",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,04,24)),
+                new AnalogyChangeLog("V4.7.0 - [Data Provider] Add Serial Port data provider. #968 ",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,04,12)),
+                new AnalogyChangeLog("V4.7.0 - Remove redundant Title text #915",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,03,20)),
+                new AnalogyChangeLog("V4.6.4 - Add 'Open in new tab' option to load files from the files list #914",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,03,20)),
+                new AnalogyChangeLog("V4.6.4 - Error during multiple loadings #913",AnalogChangeLogType.Bug,"Lior Banai",new DateTime(2021,03,20)),
+                new AnalogyChangeLog("V4.6.4 - [UI] fix file sorting in main view. #902",AnalogChangeLogType.Bug,"Lior Banai",new DateTime(2021,03,15)),
+                new AnalogyChangeLog("V4.6.3 - Add missing DevExpress DevExpress.Office.v19.1.Core.dll #900",AnalogChangeLogType.Bug,"Lior Banai",new DateTime(2021,03,12)),
+                new AnalogyChangeLog("V4.6.3 - [UI - Detailed message] add rich text box #899",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,03,12)),
+                new AnalogyChangeLog("V4.6.3 - [UI] Save user layout changes Automatically #890",AnalogChangeLogType.Bug,"Lior Banai",new DateTime(2021,02,25)),
+                new AnalogyChangeLog("V4.6.2 - [json visualizer] string not in the correct format Error #886",AnalogChangeLogType.Bug,"Lior Banai",new DateTime(2021,02,24)),
+                new AnalogyChangeLog("V4.6.1 - [UI] Fix high res monitors rendering #879",AnalogChangeLogType.Bug,"Lior Banai",new DateTime(2021,02,18)),
+                new AnalogyChangeLog("V4.6.1 - Fix floating panels sizes #878",AnalogChangeLogType.Bug,"Lior Banai",new DateTime(2021,02,18)),
+                new AnalogyChangeLog("V4.6.0 - Enable gRPC on .NET framework version #874",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,02,13)),
+                new AnalogyChangeLog("V4.6.0 - Warn if no File open filter is set for an offline data provider #871",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,02,13)),
+                new AnalogyChangeLog("V4.6.0 - Add registry settings override. #872",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,02,13)),
+                new AnalogyChangeLog("V4.6.0 - [UI] Add additional skins #873",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,02,13)),
+                new AnalogyChangeLog("V4.6.0 - [Data providers] Add custom data provider notification mechanism #516",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,02,12)),
+                new AnalogyChangeLog("V4.6.0 - [UI] Do not load subfolders when root folder/drive is selected #868",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,02,09)),
+                new AnalogyChangeLog("V4.6.0 - [UI] Refactor download/update data providers UI #867",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,02,06)),
+                new AnalogyChangeLog("V4.6.0 - [UI] Refactor data provider lists #866",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,02,06)),
+                new AnalogyChangeLog("V4.6.0 - [UI] Refactor settings form to use hamburger like UI #865",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,02,06)),
+                new AnalogyChangeLog("V4.6.0 - [UI] Improve settings and data Location wording #864",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,02,06)),
+                new AnalogyChangeLog("V4.6.0 - [UI] Reduce filtering panel space size #838",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,01,29)),
+                new AnalogyChangeLog("V4.5.0 - Add Portable Version #822",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,01,23)),
+                new AnalogyChangeLog("V4.5.0 - [UI] Add HTML/Markdown text rendering support in message details #813",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,01,13)),
+                new AnalogyChangeLog("V4.4.1 - Add Reset all setting to their default button #812",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,01,09)),
+                new AnalogyChangeLog("V4.4.1 - [UI] Make all elements in the log windows floatable #781",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,01,02)),
+                new AnalogyChangeLog("V4.4.1 - Add button to go selected message in grid #729",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,01,02)),
+                new AnalogyChangeLog("V4.4.1 - Add Import button to import existing logs into already loaded file #787",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,01,01)),
+                new AnalogyChangeLog("V4.4.1 - [Tool tips] Enable Data providers to supply their own text #737",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2021,01,01)),
+                new AnalogyChangeLog("V4.4.1 - Add application restart button when enabling/disabling data providers #778",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2020,12,25)),
+                new AnalogyChangeLog("V4.4.1 - Add option to show specific log levels (like Error,Warning...) regardless active filtering #784",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2020,12,25)),
+                new AnalogyChangeLog("V4.4.1 - [UI] Enable DPI awareness #783",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2020,12,20)),
+                new AnalogyChangeLog("V4.4.1 - [Plotting] add option to split charts to different panes #776",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2020,12,16)),
+                new AnalogyChangeLog("V4.4.1 - [Plotting] Enable toggle on/off series #775",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2020,12,16)),
+                new AnalogyChangeLog("V4.4.0 - [Real time] add option to change refresh rate [default 1 second] #768",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2020,12,12)),
+                new AnalogyChangeLog("V4.4.0 - [Feature] Data plotting API #682",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2020,12,12)),
+                new AnalogyChangeLog("V4.4.0 - [Charting] Enable real time charting #769",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2020,12,12)),
+                new AnalogyChangeLog("V4.4.0 - Add Application's settings overrider by data provider (e.g disable auto update, etc) #731",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2020,12,12)),
+                new AnalogyChangeLog("V4.4.0 - Add additional probing locations provided by data providers #732",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2020,12,11)),
+                new AnalogyChangeLog("V4.4.0 - Add option to hide irrelevant columns #734",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2020,12,11)),
+                new AnalogyChangeLog("V4.4.0 - [gRPC] Add self hosting gRPC server #735",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2020,12,07)),
+                new AnalogyChangeLog("V4.4.0 - [Debug] Reflect internal log changes in real time #736",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2020,12,06)),
+                new AnalogyChangeLog("V4.4.0 - Remove Debugger.Launch from open file association path #691",AnalogChangeLogType.Bug,"Lior Banai",new DateTime(2020,11,28)),
+                new AnalogyChangeLog("V4.4.0 - Move User Statistics outside of user settings #689",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2020,11,28)),
+                new AnalogyChangeLog("V4.4.0 - Auto restart real time provider selection is broken #688",AnalogChangeLogType.Bug,"Lior Banai",new DateTime(2020,11,28)),
+                new AnalogyChangeLog("V4.4.0 - [Feature] Restore Extensions code (User control) #687",AnalogChangeLogType.Feature,"Lior Banai",new DateTime(2020,11,28)),
                 new AnalogyChangeLog("V4.3.2 - add https://github.com/Analogy-LogViewer/Analogy.LogViewer.GitHubActionLogs to the download Window #686",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2020,11,27)),
                 new AnalogyChangeLog("V4.3.2 - Fix charts UI issues (winforms datagrid) #685",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2020,11,27)),
                 new AnalogyChangeLog("V4.3.2 - Remove grouping tab as it has low value #684",AnalogChangeLogType.Improvement,"Lior Banai",new DateTime(2020,11,27)),
