@@ -1,8 +1,9 @@
 ﻿using Newtonsoft.Json;
 using System;
+using System.Collections.Generic;
 using System.Drawing;
-using System.Net;
 using System.Windows.Forms;
+using Analogy.Interfaces;
 
 namespace Analogy.DataTypes
 {
@@ -23,6 +24,7 @@ namespace Analogy.DataTypes
 
         public override string ToString() => $"{Name} ({ID})";
     }
+
     [Serializable]
     public class AnalogyPositionState
     {
@@ -30,6 +32,7 @@ namespace Analogy.DataTypes
         public Point Location { get; set; }
         public Size Size { get; set; }
         public bool RememberLastPosition { get; set; }
+
         public AnalogyPositionState()
         {
             WindowState = FormWindowState.Maximized;
@@ -44,6 +47,7 @@ namespace Analogy.DataTypes
         }
 
     }
+
     [Serializable]
     public class FilterCriteriaUIOption
     {
@@ -59,6 +63,7 @@ namespace Analogy.DataTypes
             CheckMember = checkMember;
         }
     }
+
     public enum DataSourceType
     {
         Client,
@@ -66,6 +71,7 @@ namespace Analogy.DataTypes
         NetworkPath,
         LocalPath
     }
+
     [Serializable]
     public class DataSource
     {
@@ -73,6 +79,7 @@ namespace Analogy.DataTypes
         public string Path { get; set; }
         public DataSourceType Type { get; set; }
         public string DisplayName { get; set; }
+
         public DataSource(string hostNameOrIp, string path, string displayName, DataSourceType type)
         {
             HostNameOrIP = hostNameOrIp ?? throw new ArgumentNullException(nameof(hostNameOrIp));
@@ -99,74 +106,141 @@ namespace Analogy.DataTypes
         [JsonIgnore] public Font UIFont => new Font(FontName, FontSize, FontStyle.Regular, GraphicsUnit.Point);
         public string MenuFontName { get; set; }
         public float MenuFontSize { get; set; }
-        [JsonIgnore] public Font MenuFont => new Font(MenuFontName, MenuFontSize, FontStyle.Regular, GraphicsUnit.Point);
+
+        [JsonIgnore]
+        public Font MenuFont => new Font(MenuFontName, MenuFontSize, FontStyle.Regular, GraphicsUnit.Point);
+
         public FontSettings()
         {
-            FontSelectionType = FontSelectionType.Normal;
-            MenuFontSelectionType = FontSelectionType.Normal;
             GridFontSize = 8.5f;
-            FontName = "Tahoma";
-            FontSize = 8.25f;
-            MenuFontName = "Segoe UI";
-            MenuFontSize = 12f;
+            SetFontSelectionType(FontSelectionType.Default);
+            SetMenuFontSelectionType(FontSelectionType.Normal);
         }
 
         public void SetFontSelectionType(FontSelectionType mode)
         {
             FontSelectionType = mode;
+            FontName = "Tahoma";
             switch (mode)
             {
-                case FontSelectionType.Large:
-                    FontName = "Tahoma";
-                    FontSize = 10f;
-                    break;
-                case FontSelectionType.VeryLarge:
-                    FontName = "Tahoma";
-                    FontSize = 14f;
+                case FontSelectionType.Default:
+                    FontSize = 8.25f;
                     break;
                 case FontSelectionType.Normal:
-                case FontSelectionType.Manual:
+                    FontSize = 10f;
+                    break;
+                case FontSelectionType.Large:
+                    FontSize = 12f;
+                    break;
+                case FontSelectionType.VeryLarge:
+                    FontSize = 14f;
+                    break;
                 default:
-                    FontName = "Tahoma";
                     FontSize = 8.25f;
                     break;
             }
         }
+
+        public Font GetFontType(FontSelectionType mode)
+        {
+          var fontName = "Tahoma";
+          float fontSize;
+            switch (mode)
+            {
+                case FontSelectionType.Default:
+                    fontSize = 8.25f;
+                    break;
+                case FontSelectionType.Normal:
+                    fontSize = 10f;
+                    break;
+                case FontSelectionType.Large:
+                    fontSize = 12f;
+                    break;
+                case FontSelectionType.VeryLarge:
+                    fontSize = 14f;
+                    break;
+                default:
+                    fontSize = 8.25f;
+                    break;
+            }
+            return new Font(fontName, fontSize, FontStyle.Regular, GraphicsUnit.Point);
+
+        }
         public void SetMenuFontSelectionType(FontSelectionType mode)
         {
             MenuFontSelectionType = mode;
+            MenuFontName = "Segoe UI";
             switch (mode)
             {
+                case FontSelectionType.Default:
+                    MenuFontSize = 12f;
+                    break;
+                case FontSelectionType.Normal:
+                    MenuFontSize = 10f;
+                    break;
                 case FontSelectionType.Large:
-                    MenuFontName = "Segoe UI";
                     MenuFontSize = 14f;
                     break;
                 case FontSelectionType.VeryLarge:
-                    MenuFontName = "Segoe UI";
                     MenuFontSize = 16f;
                     break;
-                case FontSelectionType.Normal:
-                case FontSelectionType.Manual:
                 default:
-                    MenuFontName = "Segoe UI";
                     MenuFontSize = 12f;
                     break;
             }
         }
-    }
-    public class MyWebClient : WebClient
-    {
-        /// <summary>
-        ///     Response Uri after any redirects.
-        /// </summary>
-        public Uri ResponseUri;
 
-        /// <inheritdoc />
-        protected override WebResponse GetWebResponse(WebRequest request, IAsyncResult result)
+        public Font GetMenuFont(FontSelectionType mode)
         {
-            WebResponse webResponse = base.GetWebResponse(request, result);
-            ResponseUri = webResponse.ResponseUri;
-            return webResponse;
+            var menuFontName = "Segoe UI";
+            float menuFontSize;
+            switch (mode)
+            {
+                case FontSelectionType.Default:
+                    menuFontSize = 12f;
+                    break;
+                case FontSelectionType.Normal:
+                    menuFontSize = 10f;
+                    break;
+                case FontSelectionType.Large:
+                    menuFontSize = 14f;
+                    break;
+                case FontSelectionType.VeryLarge:
+                    menuFontSize = 16f;
+                    break;
+                default:
+                    menuFontSize = 12f;
+                    break;
+            }
+            return new Font(menuFontName, menuFontSize, FontStyle.Regular, GraphicsUnit.Point);
         }
+    }
+
+    [Serializable]
+    public class FilteringExclusion
+    {
+        public List<string> ExcludeTexts { get; set; }
+        public List<string> ExcludeSources { get; set; }
+        public List<string> ExcludeModules { get; set; }
+        public Dictionary<string, bool> ExcludeLogLevels { get; set; }
+
+
+        public FilteringExclusion()
+        {
+            ExcludeTexts = new List<string>();
+            ExcludeModules = new List<string>();
+            ExcludeSources = new List<string>();
+            ExcludeLogLevels = new Dictionary<string, bool>();
+            foreach (string value in Utils.LogLevels)
+            {
+                ExcludeLogLevels.Add(value, false);
+            }
+        }
+
+        public bool IsLogLevelExcluded(AnalogyLogLevel level) => IsLogLevelExcluded(level.ToString());
+        public bool IsLogLevelExcluded(string level) => ExcludeLogLevels.ContainsKey(level) && ExcludeLogLevels[level];
+
+        public void SetLogLevelExclusion(string level, bool exclude)=> ExcludeLogLevels[level] = exclude;
+        
     }
 }
