@@ -101,7 +101,7 @@ namespace Analogy
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public string GetSqlExpression()
+        public string GetSqlExpression(bool orLogLevel)
         {
 
             StringBuilder sqlString = new StringBuilder();
@@ -229,10 +229,11 @@ namespace Analogy
                 sqlString.Append(")");
             }
 
+            string andOr = orLogLevel ? "or" : "and";
             if (Levels != null && Levels.Any())
             {
                 string sTemp = string.Join(",", Levels.Select(l => $"'{l}'"));
-                sqlString.Append(" and Level in (" + sTemp + ")");
+                sqlString.Append($" {andOr} Level in (" + sTemp + ")");
             }
 
             string dateFilter = $" AND (Date >= '{NewerThan}' and Date <= '{OlderThan}')";
