@@ -1,11 +1,11 @@
-﻿using Analogy.Managers;
+﻿using Analogy.DataTypes;
+using Analogy.Managers;
 using Analogy.UserControls;
 using DevExpress.XtraEditors;
 using System;
 using System.IO;
 using System.Linq;
 using System.Windows.Forms;
-using Analogy.DataTypes;
 
 namespace Analogy.Forms
 {
@@ -19,23 +19,23 @@ namespace Analogy.Forms
         private void ComponentDownloadsForm_Load(object sender, EventArgs e)
         {
             Icon = UserSettingsManager.UserSettings.GetIcon();
-           // UserControl first = null;
+            // UserControl first = null;
             foreach (FactoryContainer factory in FactoriesManager.Instance.Factories)
             {
                 ComponentDownloadInformationUC uc = new ComponentDownloadInformationUC(factory);
                 xtraTabPage1.Controls.Add(uc);
                 uc.Dock = DockStyle.Top;
-              //  first ??= uc;
+                //  first ??= uc;
 
             }
 
-            var assemblies = FactoriesManager.Instance.Factories.Select(f=>Path.GetFileName(f.AssemblyFullPath)).ToList();
-            var supported = UserSettingsManager.UserSettings.SupportedDataProviders;
-            var notInstalled = supported.Where(s => !assemblies.Contains(s.AssemblyFileName,StringComparison.InvariantCultureIgnoreCase)).ToList();
-          
+            var assemblies = FactoriesManager.Instance.Factories.Select(f => Path.GetFileName(f.AssemblyFullPath)).ToList();
+            var supported = UpdateManager.Instance.SupportedDataProviders;
+            var notInstalled = supported.Where(s => !assemblies.Contains(s.AssemblyFileName, StringComparison.InvariantCultureIgnoreCase)).ToList();
+
             foreach (var dp in notInstalled)
             {
-                MissingDownloadInformation missing = new MissingDownloadInformation(dp.FactoryId,dp.Name,dp.RepositoryURL);
+                MissingDownloadInformation missing = new MissingDownloadInformation(dp.FactoryId, dp.Name, dp.RepositoryURL,dp.Image);
                 ComponentDownloadInformationUC uc = new ComponentDownloadInformationUC(missing);
                 xtraTabPage2.Controls.Add(uc);
                 uc.Dock = DockStyle.Top;
