@@ -1,10 +1,10 @@
-﻿using DevExpress.XtraEditors;
+﻿using Analogy.DataTypes;
+using DevExpress.XtraEditors;
+using DevExpress.XtraEditors.Controls;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Windows.Forms;
-using Analogy.DataTypes;
-using DevExpress.XtraEditors.Controls;
 
 namespace Analogy.Forms
 {
@@ -22,7 +22,7 @@ namespace Analogy.Forms
             // This line of code is generated by Data Source Configuration Wizard
         }
 
-        private void FirstTimeRunForm_Load(object sender, EventArgs e)
+        private async void FirstTimeRunForm_Load(object sender, EventArgs e)
         {
             if (DesignMode)
             {
@@ -30,6 +30,17 @@ namespace Analogy.Forms
             }
 
             Icon = UserSettingsManager.UserSettings.GetIcon();
+            await FactoriesManager.Instance.InitializeBuiltInFactories();
+            await FactoriesManager.Instance.AddExternalDataSources();
+            chkLstDataProviderStatus.CustomizeItem += (s, e) =>
+            {
+                FactoryCheckItem bind = (FactoryCheckItem)e.Value;
+                e.TemplatedItem.Elements[0].ImageOptions.Image = bind.Image;
+                e.TemplatedItem.Elements[1].Text = $"{bind.Name} (id:{bind.ID})";
+                e.TemplatedItem.Elements[2].Text = bind.Description;
+            };
+            chkLstDataProviderStatus.ItemCheck += (s, e) => SaveSettings();
+
             foreach (var setting in Settings.FactoriesOrder)
             {
                 FactorySettings factory = Settings.GetFactorySetting(setting);
@@ -63,6 +74,7 @@ namespace Analogy.Forms
 
         private void AcceptAndClose(object sender, EventArgs e)
         {
+            Settings.IsFirstRun = false;
             Close();
         }
 
@@ -84,9 +96,5 @@ namespace Analogy.Forms
             }
         }
 
-        private void chkLstDataProviderStatus_ItemCheck(object sender, DevExpress.XtraEditors.Controls.ItemCheckEventArgs e)
-        {
-            SaveSettings();
-        }
     }
 }
