@@ -10,6 +10,7 @@ namespace Analogy.Forms
         public WhatsNewForm()
         {
             InitializeComponent();
+            EnableAcrylicAccent = false;
         }
 
         private void WhatsNewForm_Load(object sender, EventArgs e)
@@ -18,9 +19,8 @@ namespace Analogy.Forms
             {
                 return;
             }
-
             Icon = UserSettingsManager.UserSettings.GetIcon();
-            WhatIsNew4_3_2 uc = new WhatIsNew4_3_2 { Name = "V4.3.2"};
+            WhatIsNew4_7_0 uc = new WhatIsNew4_7_0 { Name = "V4.7.0" };
             fluentDesignFormContainer1.Controls.Add(uc);
             uc.Dock = DockStyle.Fill;
             uc.BringToFront();
@@ -121,6 +121,72 @@ namespace Analogy.Forms
                 uc.BringToFront();
             }
             SetActive("V4.3.2");
+        }
+
+        private void e440_Click(object sender, EventArgs e)
+        {
+            if (!fluentDesignFormContainer1.Controls.ContainsKey("V4.4.0"))
+            {
+                WhatIsNew4_4_0 uc = new WhatIsNew4_4_0();
+                uc.Name = "V4.4.0";
+                fluentDesignFormContainer1.Controls.Add(uc);
+                uc.Dock = DockStyle.Fill;
+                uc.BringToFront();
+            }
+            SetActive("V4.4.0");
+        }
+
+        private void e441_Click(object sender, EventArgs e)
+        {
+
+            if (!fluentDesignFormContainer1.Controls.ContainsKey("V4.4.1"))
+            {
+                WhatIsNew4_4_1 uc = new WhatIsNew4_4_1();
+                uc.Name = "V4.4.1";
+                fluentDesignFormContainer1.Controls.Add(uc);
+                uc.Dock = DockStyle.Fill;
+                uc.BringToFront();
+            }
+            SetActive("V4.4.1");
+        }
+
+        private void e450_Click(object sender, EventArgs e)
+        {
+            if (!fluentDesignFormContainer1.Controls.ContainsKey("V4.5.0"))
+            {
+                WhatIsNew4_5_0 uc = new WhatIsNew4_5_0();
+                uc.Name = "V4.5.0";
+                fluentDesignFormContainer1.Controls.Add(uc);
+                uc.Dock = DockStyle.Fill;
+                uc.BringToFront();
+            }
+            SetActive("V4.5.0");
+        }
+
+        private void e460_Click(object sender, EventArgs e)
+        {
+            if (!fluentDesignFormContainer1.Controls.ContainsKey("V4.6.0"))
+            {
+                WhatIsNew4_6_0 uc = new WhatIsNew4_6_0();
+                uc.Name = "V4.6.0";
+                fluentDesignFormContainer1.Controls.Add(uc);
+                uc.Dock = DockStyle.Fill;
+                uc.BringToFront();
+            }
+            SetActive("V4.6.0");
+        }
+
+        private void e470_Click(object sender, EventArgs e)
+        {
+            if (!fluentDesignFormContainer1.Controls.ContainsKey("V4.7.0"))
+            {
+                WhatIsNew4_7_0 uc = new WhatIsNew4_7_0();
+                uc.Name = "V4.7.0";
+                fluentDesignFormContainer1.Controls.Add(uc);
+                uc.Dock = DockStyle.Fill;
+                uc.BringToFront();
+            }
+            SetActive("V4.7.0");
         }
     }
 }
