@@ -6,6 +6,7 @@ using System.IO;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using System.Windows.Forms;
 using Analogy.Forms;
 
 namespace Analogy.Managers
@@ -170,9 +171,9 @@ namespace Analogy.Managers
             err.Level = AnalogyLogLevel.Critical;
             OnNewMessage?.Invoke(this, (err, source));
         }
-        public void Show(MainForm mainForm)
+        public void Show(Form mainForm)
         {
-            XtraFormLogGrid msg = new XtraFormLogGrid(messages, "Analogy");
+            XtraFormLogGrid msg = new XtraFormLogGrid(messages, "Analogy",true);
             msg.Show(mainForm);
         }
 
