@@ -1,11 +1,11 @@
-﻿using System;
+﻿using Analogy.DataTypes;
+using Analogy.Interfaces;
+using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Linq;
 using System.Threading;
 using System.Windows.Forms;
-using Analogy.DataTypes;
-using Analogy.Interfaces;
 
 namespace Analogy
 {
@@ -14,7 +14,7 @@ namespace Analogy
         //public List<string> CurrentColumns { get; set; }
         private static ManualResetEvent columnAdderSync = new ManualResetEvent(false);
         public ReaderWriterLockSlim columnsLockSlim = new ReaderWriterLockSlim(LockRecursionPolicy.SupportsRecursion);
-        private readonly UCLogs owner;
+        private readonly UserControl owner;
         public event EventHandler<AnalogyClearedHistoryEventArgs> OnHistoryCleared;
         public event EventHandler<AnalogyPagingChanged> OnPageChanged;
         public ReaderWriterLockSlim lockSlim = new ReaderWriterLockSlim(LockRecursionPolicy.SupportsRecursion);
@@ -42,7 +42,7 @@ namespace Analogy
 
         public static int TotalMissedMessages => _totalMissedMessages;
 
-        public PagingManager(UCLogs owner)
+        public PagingManager(UserControl owner)
         {
             //CurrentColumns = new List<string>();
             this.owner = owner;
@@ -96,7 +96,15 @@ namespace Analogy
         public DataRow AppendMessage(AnalogyLogMessage message, string dataSource)
         {
             var table = pages.Last();
-            allMessages.Add(message);
+            try
+            {
+                lockSlim.EnterWriteLock();
+                allMessages.Add(message);
+            }
+            finally
+            {
+                lockSlim.ExitWriteLock();
+            }
             if (table.Rows.Count + 1 > pageSize)
             {
                 table = Utils.DataTableConstructor();
@@ -104,15 +112,15 @@ namespace Analogy
                 var pageStartRowIndex = (pages.Count - 1) * pageSize;
                 OnPageChanged?.Invoke(this, new AnalogyPagingChanged(new AnalogyPageInformation(table, pages.Count, pageStartRowIndex)));
             }
+
+            if (message.AdditionalInformation != null && message.AdditionalInformation.Any() && Settings.CheckAdditionalInformation)
+            {
+                AddExtraColumnsIfNeeded(table, message);
+            }
             try
             {
-                if (message.AdditionalInformation != null && message.AdditionalInformation.Any() && Settings.CheckAdditionalInformation)
-                {
-                    AddExtraColumnsIfNeeded(table, message);
-                }
-
                 lockSlim.EnterWriteLock();
-                DataRow dtr = Utils.CreateRow(table, message, dataSource,Settings.CheckAdditionalInformation);
+                DataRow dtr = Utils.CreateRow(table, message, dataSource, Settings.CheckAdditionalInformation);
                 table.Rows.Add(dtr);
                 return dtr;
 
@@ -125,27 +133,38 @@ namespace Analogy
 
         public List<(DataRow, AnalogyLogMessage)> AppendMessages(List<AnalogyLogMessage> messages, string dataSource)
         {
+
             var table = pages.Last();
             var countInsideTable = table.Rows.Count;
             List<(DataRow row, AnalogyLogMessage message)> rows = new List<(DataRow row, AnalogyLogMessage message)>(messages.Count);
             foreach (var message in messages)
             {
-
                 if (message.Level == AnalogyLogLevel.None)
                 {
                     continue; //ignore those messages
                 }
-
-                allMessages.Add(message);
+                try
+                {
+                    lockSlim.EnterWriteLock();
+                    allMessages.Add(message);
+                }
+                finally
+                {
+                    lockSlim.ExitWriteLock();
+                }
                 if (countInsideTable + 1 > pageSize)
                 {
                     table = Utils.DataTableConstructor();
                     pages.Add(table);
                     countInsideTable = 0;
                     var pageStartRowIndex = (pages.Count - 1) * pageSize;
-                    OnPageChanged?.Invoke(this, new AnalogyPagingChanged(new AnalogyPageInformation(table, pages.Count, pageStartRowIndex)));
+                    OnPageChanged?.Invoke(this,
+                        new AnalogyPagingChanged(new AnalogyPageInformation(table, pages.Count,
+                            pageStartRowIndex)));
                 }
-                if (message.AdditionalInformation != null && message.AdditionalInformation.Any() && Settings.CheckAdditionalInformation)
+
+                if (message.AdditionalInformation != null && message.AdditionalInformation.Any() &&
+                    Settings.CheckAdditionalInformation)
                 {
                     AddExtraColumnsIfNeeded(table, message);
                 }
@@ -178,26 +197,45 @@ namespace Analogy
                     {
                         if (!owner.InvokeRequired)
                         {
-                            columnsLockSlim.EnterWriteLock();
-                            if (!currentTable.Columns.Contains(info.Key))
-                            {
-                                table.Columns.Add(info.Key);
-                            }
-
-                            columnsLockSlim.ExitWriteLock();
-                        }
-                        else
-                        {
-                            owner.BeginInvoke(new MethodInvoker(() =>
+                            try
                             {
                                 columnsLockSlim.EnterWriteLock();
                                 if (!currentTable.Columns.Contains(info.Key))
                                 {
                                     table.Columns.Add(info.Key);
-                                    columnAdderSync.Set();
-
                                 }
+                            }
+                            finally
+                            {
                                 columnsLockSlim.ExitWriteLock();
+                            }
+
+
+
+                        }
+                        else
+                        {
+                            owner.BeginInvoke(new MethodInvoker(() =>
+                            {
+                                try
+                                {
+
+                                    columnsLockSlim.EnterWriteLock();
+                                    if (!currentTable.Columns.Contains(info.Key))
+                                    {
+                                        columnsLockSlim.EnterWriteLock();
+                                        if (!currentTable.Columns.Contains(info.Key))
+                                        {
+                                            table.Columns.Add(info.Key);
+                                            columnAdderSync.Set();
+                                        }
+
+                                    }
+                                }
+                                finally
+                                {
+                                    columnsLockSlim.ExitWriteLock();
+                                }
                             }));
                             columnAdderSync.WaitOne();
                             columnAdderSync.Reset();
@@ -257,10 +295,16 @@ namespace Analogy
 
         public List<AnalogyLogMessage> GetAllMessages()
         {
-            lockSlim.EnterReadLock();
-            var items = allMessages.ToList();
-            lockSlim.ExitReadLock();
-            return items;
+            try
+            {
+                lockSlim.EnterReadLock();
+                var items = allMessages.ToList();
+                return items;
+            }
+            finally
+            {
+                lockSlim.ExitReadLock();
+            }
         }
 
         public DataTable CurrentPage()
@@ -283,6 +327,22 @@ namespace Analogy
         public void IncrementTotalMissedMessages()
         {
             Interlocked.Increment(ref _totalMissedMessages);
+        }
+
+        public void UpdateOffsets()
+        {
+            lockSlim.EnterWriteLock();
+            foreach (DataTable dataTable in pages)
+            {
+                foreach (DataRow dataTableRow in dataTable.Rows)
+                {
+                    dataTableRow.BeginEdit();
+                    AnalogyLogMessage m = (AnalogyLogMessage)dataTableRow["Object"];
+                    dataTableRow["Date"] = Utils.GetOffsetTime(m.Date);
+                    dataTableRow.EndEdit();
+                }
+            }
+            lockSlim.ExitWriteLock();
         }
     }
 
