@@ -1,16 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Text;
-using System.Linq;
-using System.Threading.Tasks;
-using System.Windows.Forms;
+﻿using Analogy.Interfaces;
 using Analogy.Tools.JsonViewer;
-using DevExpress.XtraEditors;
-using Analogy.Interfaces;
 using Newtonsoft.Json;
+using System;
+using System.Windows.Forms;
 
 namespace Analogy.Tools
 {
@@ -19,15 +11,17 @@ namespace Analogy.Tools
         private UserSettingsManager settings => UserSettingsManager.UserSettings;
         private JsonTreeView _jsonTreeView;
         private AnalogyLogMessage message;
+        private readonly bool _useRawField;
 
         public JsonViewerForm()
         {
             InitializeComponent();
         }
 
-        public JsonViewerForm(AnalogyLogMessage message):this()
+        public JsonViewerForm(AnalogyLogMessage message) : this()
         {
             this.message = message;
+            _useRawField = message.RawTextType == AnalogyRowTextType.JSON;
         }
 
         private void JsonViewerForm_Load(object sender, EventArgs e)
@@ -38,13 +32,14 @@ namespace Analogy.Tools
             }
 
             Icon = settings.GetIcon();
-            _jsonTreeView = new JsonTreeView();
-            splitContainerControl1.Panel2.Controls.Add(_jsonTreeView);
+            _jsonTreeView = new JsonTreeView(); 
+            _jsonTreeView.OnNodeChanged += (s, e) => meSelected.Text = e;
+            splitContainerControl2.Panel1.Controls.Add(_jsonTreeView);
             _jsonTreeView.Dock = DockStyle.Fill;
             if (message != null)
             {
-                memoEdit1.Text = message.Text;
-                var json = ExtractJsonObject(message.Text);
+                memoEdit1.Text = _useRawField ? message.RawText : message.Text;
+                var json =Utils.ExtractJsonObject(_useRawField ? message.RawText : message.Text);
                 if (!string.IsNullOrEmpty(json))
                 {
                     _jsonTreeView.ShowJson(json);
@@ -54,28 +49,7 @@ namespace Analogy.Tools
 
 
         }
-
-        public string ExtractJsonObject(string mixedString)
-        {
-            for (var i = mixedString.IndexOf('{'); i > -1; i = mixedString.IndexOf('{', i + 1))
-            {
-                for (var j = mixedString.LastIndexOf('}'); j > -1; j = mixedString.LastIndexOf("}", j - 1))
-                {
-                    var jsonProbe = mixedString.Substring(i, j - i + 1);
-                    try
-                    {
-                        var valid = JsonConvert.DeserializeObject(jsonProbe);
-                        return jsonProbe;
-                    }
-                    catch
-                    {
-                    }
-                }
-            }
-
-            return null;
-        }
-
+        
         private void sbtnLoad_Click(object sender, EventArgs e)
         {
             _jsonTreeView.ShowJson(memoEdit1.Text);
