@@ -29,6 +29,10 @@
         private void InitializeComponent()
         {
             System.ComponentModel.ComponentResourceManager resources = new System.ComponentModel.ComponentResourceManager(typeof(MessageDetailsUC));
+            DevExpress.Utils.SuperToolTip superToolTip3 = new DevExpress.Utils.SuperToolTip();
+            DevExpress.Utils.ToolTipItem toolTipItem3 = new DevExpress.Utils.ToolTipItem();
+            DevExpress.Utils.SuperToolTip superToolTip4 = new DevExpress.Utils.SuperToolTip();
+            DevExpress.Utils.ToolTipItem toolTipItem4 = new DevExpress.Utils.ToolTipItem();
             this.btnNext = new DevExpress.XtraEditors.SimpleButton();
             this.btnPrev = new DevExpress.XtraEditors.SimpleButton();
             this.simpleButton1 = new DevExpress.XtraEditors.SimpleButton();
@@ -63,8 +67,13 @@
             this.txtbThreadId = new DevExpress.XtraEditors.TextEdit();
             this.xtraTabControlMessageInfo = new DevExpress.XtraTab.XtraTabControl();
             this.xtraTabPageText = new DevExpress.XtraTab.XtraTabPage();
+            this.xtraTabPageRenderedText = new DevExpress.XtraTab.XtraTabPage();
+            this.recMessageDetails = new DevExpress.XtraRichEdit.RichEditControl();
             this.xtraTabPageAdditionalInformation = new DevExpress.XtraTab.XtraTabPage();
             this.memoAdditionalInformation = new DevExpress.XtraEditors.MemoEdit();
+            this.splitContainerControl1 = new DevExpress.XtraEditors.SplitContainerControl();
+            this.splitContainerControl2 = new DevExpress.XtraEditors.SplitContainerControl();
+            this.meSelected = new DevExpress.XtraEditors.MemoEdit();
             ((System.ComponentModel.ISupportInitialize)(this.txtbDateValue.Properties)).BeginInit();
             ((System.ComponentModel.ISupportInitialize)(this.txtbDataSource.Properties)).BeginInit();
             ((System.ComponentModel.ISupportInitialize)(this.txtSourceValue.Properties)).BeginInit();
@@ -83,17 +92,33 @@
             ((System.ComponentModel.ISupportInitialize)(this.xtraTabControlMessageInfo)).BeginInit();
             this.xtraTabControlMessageInfo.SuspendLayout();
             this.xtraTabPageText.SuspendLayout();
+            this.xtraTabPageRenderedText.SuspendLayout();
             this.xtraTabPageAdditionalInformation.SuspendLayout();
             ((System.ComponentModel.ISupportInitialize)(this.memoAdditionalInformation.Properties)).BeginInit();
+            ((System.ComponentModel.ISupportInitialize)(this.splitContainerControl1)).BeginInit();
+            ((System.ComponentModel.ISupportInitialize)(this.splitContainerControl1.Panel1)).BeginInit();
+            this.splitContainerControl1.Panel1.SuspendLayout();
+            ((System.ComponentModel.ISupportInitialize)(this.splitContainerControl1.Panel2)).BeginInit();
+            this.splitContainerControl1.Panel2.SuspendLayout();
+            this.splitContainerControl1.SuspendLayout();
+            ((System.ComponentModel.ISupportInitialize)(this.splitContainerControl2)).BeginInit();
+            ((System.ComponentModel.ISupportInitialize)(this.splitContainerControl2.Panel1)).BeginInit();
+            ((System.ComponentModel.ISupportInitialize)(this.splitContainerControl2.Panel2)).BeginInit();
+            this.splitContainerControl2.Panel2.SuspendLayout();
+            this.splitContainerControl2.SuspendLayout();
+            ((System.ComponentModel.ISupportInitialize)(this.meSelected.Properties)).BeginInit();
             this.SuspendLayout();
             // 
             // btnNext
             // 
             this.btnNext.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Right)));
             this.btnNext.ImageOptions.Image = ((System.Drawing.Image)(resources.GetObject("btnNext.ImageOptions.Image")));
-            this.btnNext.Location = new System.Drawing.Point(846, 2);
+            this.btnNext.Location = new System.Drawing.Point(555, 1);
             this.btnNext.Name = "btnNext";
             this.btnNext.Size = new System.Drawing.Size(24, 26);
+            toolTipItem3.Text = "Press right arrow to load next message";
+            superToolTip3.Items.Add(toolTipItem3);
+            this.btnNext.SuperTip = superToolTip3;
             this.btnNext.TabIndex = 38;
             this.btnNext.Click += new System.EventHandler(this.btnNext_Click);
             // 
@@ -101,9 +126,12 @@
             // 
             this.btnPrev.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Right)));
             this.btnPrev.ImageOptions.Image = ((System.Drawing.Image)(resources.GetObject("btnPrev.ImageOptions.Image")));
-            this.btnPrev.Location = new System.Drawing.Point(816, 2);
+            this.btnPrev.Location = new System.Drawing.Point(525, 1);
             this.btnPrev.Name = "btnPrev";
             this.btnPrev.Size = new System.Drawing.Size(24, 26);
+            toolTipItem4.Text = "Press Left arrow to load previous message";
+            superToolTip4.Items.Add(toolTipItem4);
+            this.btnPrev.SuperTip = superToolTip4;
             this.btnPrev.TabIndex = 39;
             this.btnPrev.Click += new System.EventHandler(this.btnPrev_Click);
             // 
@@ -119,9 +147,9 @@
             // 
             this.txtbDateValue.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left) 
             | System.Windows.Forms.AnchorStyles.Right)));
-            this.txtbDateValue.Location = new System.Drawing.Point(227, 314);
+            this.txtbDateValue.Location = new System.Drawing.Point(179, 325);
             this.txtbDateValue.Name = "txtbDateValue";
-            this.txtbDateValue.Size = new System.Drawing.Size(641, 22);
+            this.txtbDateValue.Size = new System.Drawing.Size(398, 22);
             this.txtbDateValue.TabIndex = 47;
             // 
             // labelControl1
@@ -129,9 +157,9 @@
             this.labelControl1.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
             this.labelControl1.AutoSizeMode = DevExpress.XtraEditors.LabelAutoSizeMode.None;
             this.labelControl1.BorderStyle = DevExpress.XtraEditors.Controls.BorderStyles.Flat;
-            this.labelControl1.Location = new System.Drawing.Point(6, 314);
+            this.labelControl1.Location = new System.Drawing.Point(3, 325);
             this.labelControl1.Name = "labelControl1";
-            this.labelControl1.Size = new System.Drawing.Size(216, 22);
+            this.labelControl1.Size = new System.Drawing.Size(170, 22);
             this.labelControl1.TabIndex = 48;
             this.labelControl1.Text = "Date:";
             // 
@@ -139,9 +167,9 @@
             // 
             this.txtbDataSource.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left) 
             | System.Windows.Forms.AnchorStyles.Right)));
-            this.txtbDataSource.Location = new System.Drawing.Point(226, 286);
+            this.txtbDataSource.Location = new System.Drawing.Point(178, 297);
             this.txtbDataSource.Name = "txtbDataSource";
-            this.txtbDataSource.Size = new System.Drawing.Size(641, 22);
+            this.txtbDataSource.Size = new System.Drawing.Size(398, 22);
             this.txtbDataSource.TabIndex = 47;
             // 
             // labelControl2
@@ -149,9 +177,9 @@
             this.labelControl2.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
             this.labelControl2.AutoSizeMode = DevExpress.XtraEditors.LabelAutoSizeMode.None;
             this.labelControl2.BorderStyle = DevExpress.XtraEditors.Controls.BorderStyles.Flat;
-            this.labelControl2.Location = new System.Drawing.Point(6, 286);
+            this.labelControl2.Location = new System.Drawing.Point(3, 297);
             this.labelControl2.Name = "labelControl2";
-            this.labelControl2.Size = new System.Drawing.Size(216, 22);
+            this.labelControl2.Size = new System.Drawing.Size(170, 22);
             this.labelControl2.TabIndex = 48;
             this.labelControl2.Text = "Log file / Data source:";
             // 
@@ -159,9 +187,9 @@
             // 
             this.txtSourceValue.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left) 
             | System.Windows.Forms.AnchorStyles.Right)));
-            this.txtSourceValue.Location = new System.Drawing.Point(227, 342);
+            this.txtSourceValue.Location = new System.Drawing.Point(179, 353);
             this.txtSourceValue.Name = "txtSourceValue";
-            this.txtSourceValue.Size = new System.Drawing.Size(641, 22);
+            this.txtSourceValue.Size = new System.Drawing.Size(398, 22);
             this.txtSourceValue.TabIndex = 47;
             // 
             // lblSource
@@ -169,9 +197,9 @@
             this.lblSource.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
             this.lblSource.AutoSizeMode = DevExpress.XtraEditors.LabelAutoSizeMode.None;
             this.lblSource.BorderStyle = DevExpress.XtraEditors.Controls.BorderStyles.Flat;
-            this.lblSource.Location = new System.Drawing.Point(6, 342);
+            this.lblSource.Location = new System.Drawing.Point(3, 353);
             this.lblSource.Name = "lblSource";
-            this.lblSource.Size = new System.Drawing.Size(216, 22);
+            this.lblSource.Size = new System.Drawing.Size(170, 22);
             this.lblSource.TabIndex = 48;
             this.lblSource.Text = "Source:";
             // 
@@ -179,9 +207,9 @@
             // 
             this.txtbProcessModuleName.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left) 
             | System.Windows.Forms.AnchorStyles.Right)));
-            this.txtbProcessModuleName.Location = new System.Drawing.Point(227, 370);
+            this.txtbProcessModuleName.Location = new System.Drawing.Point(179, 381);
             this.txtbProcessModuleName.Name = "txtbProcessModuleName";
-            this.txtbProcessModuleName.Size = new System.Drawing.Size(641, 22);
+            this.txtbProcessModuleName.Size = new System.Drawing.Size(398, 22);
             this.txtbProcessModuleName.TabIndex = 47;
             // 
             // lblProcess
@@ -189,9 +217,9 @@
             this.lblProcess.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
             this.lblProcess.AutoSizeMode = DevExpress.XtraEditors.LabelAutoSizeMode.None;
             this.lblProcess.BorderStyle = DevExpress.XtraEditors.Controls.BorderStyles.Flat;
-            this.lblProcess.Location = new System.Drawing.Point(6, 370);
+            this.lblProcess.Location = new System.Drawing.Point(3, 381);
             this.lblProcess.Name = "lblProcess";
-            this.lblProcess.Size = new System.Drawing.Size(216, 22);
+            this.lblProcess.Size = new System.Drawing.Size(170, 22);
             this.lblProcess.TabIndex = 48;
             this.lblProcess.Text = "Process /Module:";
             // 
@@ -199,18 +227,18 @@
             // 
             this.txtbLevelValue.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left) 
             | System.Windows.Forms.AnchorStyles.Right)));
-            this.txtbLevelValue.Location = new System.Drawing.Point(227, 455);
+            this.txtbLevelValue.Location = new System.Drawing.Point(179, 466);
             this.txtbLevelValue.Name = "txtbLevelValue";
-            this.txtbLevelValue.Size = new System.Drawing.Size(641, 22);
+            this.txtbLevelValue.Size = new System.Drawing.Size(398, 22);
             this.txtbLevelValue.TabIndex = 47;
             // 
             // txtbMethod
             // 
             this.txtbMethod.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left) 
             | System.Windows.Forms.AnchorStyles.Right)));
-            this.txtbMethod.Location = new System.Drawing.Point(227, 483);
+            this.txtbMethod.Location = new System.Drawing.Point(179, 494);
             this.txtbMethod.Name = "txtbMethod";
-            this.txtbMethod.Size = new System.Drawing.Size(641, 22);
+            this.txtbMethod.Size = new System.Drawing.Size(398, 22);
             this.txtbMethod.TabIndex = 47;
             // 
             // lblLevel
@@ -218,9 +246,9 @@
             this.lblLevel.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
             this.lblLevel.AutoSizeMode = DevExpress.XtraEditors.LabelAutoSizeMode.None;
             this.lblLevel.BorderStyle = DevExpress.XtraEditors.Controls.BorderStyles.Flat;
-            this.lblLevel.Location = new System.Drawing.Point(6, 455);
+            this.lblLevel.Location = new System.Drawing.Point(3, 466);
             this.lblLevel.Name = "lblLevel";
-            this.lblLevel.Size = new System.Drawing.Size(216, 22);
+            this.lblLevel.Size = new System.Drawing.Size(170, 22);
             this.lblLevel.TabIndex = 48;
             this.lblLevel.Text = "Source:";
             // 
@@ -229,9 +257,9 @@
             this.lblMethod.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
             this.lblMethod.AutoSizeMode = DevExpress.XtraEditors.LabelAutoSizeMode.None;
             this.lblMethod.BorderStyle = DevExpress.XtraEditors.Controls.BorderStyles.Flat;
-            this.lblMethod.Location = new System.Drawing.Point(6, 483);
+            this.lblMethod.Location = new System.Drawing.Point(3, 494);
             this.lblMethod.Name = "lblMethod";
-            this.lblMethod.Size = new System.Drawing.Size(216, 22);
+            this.lblMethod.Size = new System.Drawing.Size(170, 22);
             this.lblMethod.TabIndex = 48;
             this.lblMethod.Text = "Method:";
             // 
@@ -239,9 +267,9 @@
             // 
             this.txtbFileName.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left) 
             | System.Windows.Forms.AnchorStyles.Right)));
-            this.txtbFileName.Location = new System.Drawing.Point(225, 511);
+            this.txtbFileName.Location = new System.Drawing.Point(177, 522);
             this.txtbFileName.Name = "txtbFileName";
-            this.txtbFileName.Size = new System.Drawing.Size(641, 22);
+            this.txtbFileName.Size = new System.Drawing.Size(398, 22);
             this.txtbFileName.TabIndex = 47;
             // 
             // lblSourceCodeFileName
@@ -249,9 +277,9 @@
             this.lblSourceCodeFileName.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
             this.lblSourceCodeFileName.AutoSizeMode = DevExpress.XtraEditors.LabelAutoSizeMode.None;
             this.lblSourceCodeFileName.BorderStyle = DevExpress.XtraEditors.Controls.BorderStyles.Flat;
-            this.lblSourceCodeFileName.Location = new System.Drawing.Point(4, 511);
+            this.lblSourceCodeFileName.Location = new System.Drawing.Point(1, 522);
             this.lblSourceCodeFileName.Name = "lblSourceCodeFileName";
-            this.lblSourceCodeFileName.Size = new System.Drawing.Size(216, 22);
+            this.lblSourceCodeFileName.Size = new System.Drawing.Size(170, 22);
             this.lblSourceCodeFileName.TabIndex = 48;
             this.lblSourceCodeFileName.Text = "Source Code FileName:";
             // 
@@ -259,9 +287,9 @@
             // 
             this.txtbLineNumber.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left) 
             | System.Windows.Forms.AnchorStyles.Right)));
-            this.txtbLineNumber.Location = new System.Drawing.Point(225, 539);
+            this.txtbLineNumber.Location = new System.Drawing.Point(177, 550);
             this.txtbLineNumber.Name = "txtbLineNumber";
-            this.txtbLineNumber.Size = new System.Drawing.Size(641, 22);
+            this.txtbLineNumber.Size = new System.Drawing.Size(398, 22);
             this.txtbLineNumber.TabIndex = 47;
             // 
             // lblLineNumber
@@ -269,9 +297,9 @@
             this.lblLineNumber.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
             this.lblLineNumber.AutoSizeMode = DevExpress.XtraEditors.LabelAutoSizeMode.None;
             this.lblLineNumber.BorderStyle = DevExpress.XtraEditors.Controls.BorderStyles.Flat;
-            this.lblLineNumber.Location = new System.Drawing.Point(4, 539);
+            this.lblLineNumber.Location = new System.Drawing.Point(1, 550);
             this.lblLineNumber.Name = "lblLineNumber";
-            this.lblLineNumber.Size = new System.Drawing.Size(216, 22);
+            this.lblLineNumber.Size = new System.Drawing.Size(170, 22);
             this.lblLineNumber.TabIndex = 48;
             this.lblLineNumber.Text = "Line Number:";
             // 
@@ -279,9 +307,9 @@
             // 
             this.txtbUser.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left) 
             | System.Windows.Forms.AnchorStyles.Right)));
-            this.txtbUser.Location = new System.Drawing.Point(225, 567);
+            this.txtbUser.Location = new System.Drawing.Point(177, 578);
             this.txtbUser.Name = "txtbUser";
-            this.txtbUser.Size = new System.Drawing.Size(641, 22);
+            this.txtbUser.Size = new System.Drawing.Size(398, 22);
             this.txtbUser.TabIndex = 47;
             // 
             // lblUser
@@ -289,9 +317,9 @@
             this.lblUser.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
             this.lblUser.AutoSizeMode = DevExpress.XtraEditors.LabelAutoSizeMode.None;
             this.lblUser.BorderStyle = DevExpress.XtraEditors.Controls.BorderStyles.Flat;
-            this.lblUser.Location = new System.Drawing.Point(4, 567);
+            this.lblUser.Location = new System.Drawing.Point(1, 578);
             this.lblUser.Name = "lblUser";
-            this.lblUser.Size = new System.Drawing.Size(216, 22);
+            this.lblUser.Size = new System.Drawing.Size(170, 22);
             this.lblUser.TabIndex = 48;
             this.lblUser.Text = "User:";
             // 
@@ -299,16 +327,16 @@
             // 
             this.txtID.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Left) 
             | System.Windows.Forms.AnchorStyles.Right)));
-            this.txtID.Location = new System.Drawing.Point(225, 34);
+            this.txtID.Location = new System.Drawing.Point(225, 31);
             this.txtID.Name = "txtID";
-            this.txtID.Size = new System.Drawing.Size(641, 22);
+            this.txtID.Size = new System.Drawing.Size(354, 22);
             this.txtID.TabIndex = 49;
             // 
             // labelControl3
             // 
             this.labelControl3.AutoSizeMode = DevExpress.XtraEditors.LabelAutoSizeMode.None;
             this.labelControl3.BorderStyle = DevExpress.XtraEditors.Controls.BorderStyles.Flat;
-            this.labelControl3.Location = new System.Drawing.Point(0, 34);
+            this.labelControl3.Location = new System.Drawing.Point(3, 31);
             this.labelControl3.Name = "labelControl3";
             this.labelControl3.Size = new System.Drawing.Size(216, 22);
             this.labelControl3.TabIndex = 48;
@@ -318,7 +346,7 @@
             // 
             this.labelControl4.AutoSizeMode = DevExpress.XtraEditors.LabelAutoSizeMode.None;
             this.labelControl4.BorderStyle = DevExpress.XtraEditors.Controls.BorderStyles.Flat;
-            this.labelControl4.Location = new System.Drawing.Point(0, 4);
+            this.labelControl4.Location = new System.Drawing.Point(3, 3);
             this.labelControl4.Name = "labelControl4";
             this.labelControl4.Size = new System.Drawing.Size(216, 22);
             this.labelControl4.TabIndex = 48;
@@ -328,9 +356,9 @@
             // 
             this.txtbIndex.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Left) 
             | System.Windows.Forms.AnchorStyles.Right)));
-            this.txtbIndex.Location = new System.Drawing.Point(222, 4);
+            this.txtbIndex.Location = new System.Drawing.Point(225, 3);
             this.txtbIndex.Name = "txtbIndex";
-            this.txtbIndex.Size = new System.Drawing.Size(588, 22);
+            this.txtbIndex.Size = new System.Drawing.Size(294, 22);
             this.txtbIndex.TabIndex = 49;
             // 
             // memoText
@@ -338,7 +366,7 @@
             this.memoText.Dock = System.Windows.Forms.DockStyle.Fill;
             this.memoText.Location = new System.Drawing.Point(0, 0);
             this.memoText.Name = "memoText";
-            this.memoText.Size = new System.Drawing.Size(856, 154);
+            this.memoText.Size = new System.Drawing.Size(574, 172);
             this.memoText.TabIndex = 50;
             // 
             // lblMachineName
@@ -346,9 +374,9 @@
             this.lblMachineName.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
             this.lblMachineName.AutoSizeMode = DevExpress.XtraEditors.LabelAutoSizeMode.None;
             this.lblMachineName.BorderStyle = DevExpress.XtraEditors.Controls.BorderStyles.Flat;
-            this.lblMachineName.Location = new System.Drawing.Point(7, 256);
+            this.lblMachineName.Location = new System.Drawing.Point(4, 267);
             this.lblMachineName.Name = "lblMachineName";
-            this.lblMachineName.Size = new System.Drawing.Size(216, 22);
+            this.lblMachineName.Size = new System.Drawing.Size(170, 22);
             this.lblMachineName.TabIndex = 52;
             this.lblMachineName.Text = "Machine Name:";
             // 
@@ -356,9 +384,9 @@
             // 
             this.txtbMachineName.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left) 
             | System.Windows.Forms.AnchorStyles.Right)));
-            this.txtbMachineName.Location = new System.Drawing.Point(227, 256);
+            this.txtbMachineName.Location = new System.Drawing.Point(179, 267);
             this.txtbMachineName.Name = "txtbMachineName";
-            this.txtbMachineName.Size = new System.Drawing.Size(641, 22);
+            this.txtbMachineName.Size = new System.Drawing.Size(398, 22);
             this.txtbMachineName.TabIndex = 51;
             // 
             // labelControl5
@@ -366,9 +394,9 @@
             this.labelControl5.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
             this.labelControl5.AutoSizeMode = DevExpress.XtraEditors.LabelAutoSizeMode.None;
             this.labelControl5.BorderStyle = DevExpress.XtraEditors.Controls.BorderStyles.Flat;
-            this.labelControl5.Location = new System.Drawing.Point(6, 398);
+            this.labelControl5.Location = new System.Drawing.Point(3, 409);
             this.labelControl5.Name = "labelControl5";
-            this.labelControl5.Size = new System.Drawing.Size(216, 22);
+            this.labelControl5.Size = new System.Drawing.Size(170, 22);
             this.labelControl5.TabIndex = 54;
             this.labelControl5.Text = "Process Id:";
             // 
@@ -376,9 +404,9 @@
             // 
             this.txtbProcessId.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left) 
             | System.Windows.Forms.AnchorStyles.Right)));
-            this.txtbProcessId.Location = new System.Drawing.Point(227, 398);
+            this.txtbProcessId.Location = new System.Drawing.Point(179, 409);
             this.txtbProcessId.Name = "txtbProcessId";
-            this.txtbProcessId.Size = new System.Drawing.Size(641, 22);
+            this.txtbProcessId.Size = new System.Drawing.Size(398, 22);
             this.txtbProcessId.TabIndex = 53;
             // 
             // lblThreadId
@@ -386,9 +414,9 @@
             this.lblThreadId.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
             this.lblThreadId.AutoSizeMode = DevExpress.XtraEditors.LabelAutoSizeMode.None;
             this.lblThreadId.BorderStyle = DevExpress.XtraEditors.Controls.BorderStyles.Flat;
-            this.lblThreadId.Location = new System.Drawing.Point(6, 426);
+            this.lblThreadId.Location = new System.Drawing.Point(3, 437);
             this.lblThreadId.Name = "lblThreadId";
-            this.lblThreadId.Size = new System.Drawing.Size(216, 22);
+            this.lblThreadId.Size = new System.Drawing.Size(170, 22);
             this.lblThreadId.TabIndex = 56;
             this.lblThreadId.Text = "Thread Id:";
             // 
@@ -396,9 +424,9 @@
             // 
             this.txtbThreadId.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left) 
             | System.Windows.Forms.AnchorStyles.Right)));
-            this.txtbThreadId.Location = new System.Drawing.Point(227, 426);
+            this.txtbThreadId.Location = new System.Drawing.Point(179, 437);
             this.txtbThreadId.Name = "txtbThreadId";
-            this.txtbThreadId.Size = new System.Drawing.Size(641, 22);
+            this.txtbThreadId.Size = new System.Drawing.Size(398, 22);
             this.txtbThreadId.TabIndex = 55;
             // 
             // xtraTabControlMessageInfo
@@ -406,12 +434,13 @@
             this.xtraTabControlMessageInfo.Anchor = ((System.Windows.Forms.AnchorStyles)((((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Bottom) 
             | System.Windows.Forms.AnchorStyles.Left) 
             | System.Windows.Forms.AnchorStyles.Right)));
-            this.xtraTabControlMessageInfo.Location = new System.Drawing.Point(3, 62);
+            this.xtraTabControlMessageInfo.Location = new System.Drawing.Point(3, 59);
             this.xtraTabControlMessageInfo.Name = "xtraTabControlMessageInfo";
             this.xtraTabControlMessageInfo.SelectedTabPage = this.xtraTabPageText;
-            this.xtraTabControlMessageInfo.Size = new System.Drawing.Size(863, 188);
+            this.xtraTabControlMessageInfo.Size = new System.Drawing.Size(576, 202);
             this.xtraTabControlMessageInfo.TabIndex = 57;
             this.xtraTabControlMessageInfo.TabPages.AddRange(new DevExpress.XtraTab.XtraTabPage[] {
+            this.xtraTabPageRenderedText,
             this.xtraTabPageText,
             this.xtraTabPageAdditionalInformation});
             // 
@@ -419,14 +448,31 @@
             // 
             this.xtraTabPageText.Controls.Add(this.memoText);
             this.xtraTabPageText.Name = "xtraTabPageText";
-            this.xtraTabPageText.Size = new System.Drawing.Size(856, 154);
+            this.xtraTabPageText.Size = new System.Drawing.Size(574, 172);
             this.xtraTabPageText.Text = "Message Text";
+            // 
+            // xtraTabPageRenderedText
+            // 
+            this.xtraTabPageRenderedText.Controls.Add(this.recMessageDetails);
+            this.xtraTabPageRenderedText.Name = "xtraTabPageRenderedText";
+            this.xtraTabPageRenderedText.Size = new System.Drawing.Size(574, 172);
+            this.xtraTabPageRenderedText.Text = "Formatted Text";
+            // 
+            // recMessageDetails
+            // 
+            this.recMessageDetails.ActiveViewType = DevExpress.XtraRichEdit.RichEditViewType.Simple;
+            this.recMessageDetails.Dock = System.Windows.Forms.DockStyle.Fill;
+            this.recMessageDetails.LayoutUnit = DevExpress.XtraRichEdit.DocumentLayoutUnit.Pixel;
+            this.recMessageDetails.Location = new System.Drawing.Point(0, 0);
+            this.recMessageDetails.Name = "recMessageDetails";
+            this.recMessageDetails.Size = new System.Drawing.Size(574, 172);
+            this.recMessageDetails.TabIndex = 5;
             // 
             // xtraTabPageAdditionalInformation
             // 
             this.xtraTabPageAdditionalInformation.Controls.Add(this.memoAdditionalInformation);
             this.xtraTabPageAdditionalInformation.Name = "xtraTabPageAdditionalInformation";
-            this.xtraTabPageAdditionalInformation.Size = new System.Drawing.Size(856, 154);
+            this.xtraTabPageAdditionalInformation.Size = new System.Drawing.Size(574, 172);
             this.xtraTabPageAdditionalInformation.Text = "Message\'s Additional Information";
             // 
             // memoAdditionalInformation
@@ -434,8 +480,87 @@
             this.memoAdditionalInformation.Dock = System.Windows.Forms.DockStyle.Fill;
             this.memoAdditionalInformation.Location = new System.Drawing.Point(0, 0);
             this.memoAdditionalInformation.Name = "memoAdditionalInformation";
-            this.memoAdditionalInformation.Size = new System.Drawing.Size(856, 154);
+            this.memoAdditionalInformation.Size = new System.Drawing.Size(574, 172);
             this.memoAdditionalInformation.TabIndex = 51;
+            // 
+            // splitContainerControl1
+            // 
+            this.splitContainerControl1.Dock = System.Windows.Forms.DockStyle.Fill;
+            this.splitContainerControl1.FixedPanel = DevExpress.XtraEditors.SplitFixedPanel.Panel2;
+            this.splitContainerControl1.Location = new System.Drawing.Point(0, 0);
+            this.splitContainerControl1.Name = "splitContainerControl1";
+            // 
+            // splitContainerControl1.Panel1
+            // 
+            this.splitContainerControl1.Panel1.Controls.Add(this.lblThreadId);
+            this.splitContainerControl1.Panel1.Controls.Add(this.xtraTabControlMessageInfo);
+            this.splitContainerControl1.Panel1.Controls.Add(this.txtbThreadId);
+            this.splitContainerControl1.Panel1.Controls.Add(this.labelControl4);
+            this.splitContainerControl1.Panel1.Controls.Add(this.labelControl5);
+            this.splitContainerControl1.Panel1.Controls.Add(this.txtbIndex);
+            this.splitContainerControl1.Panel1.Controls.Add(this.txtbProcessId);
+            this.splitContainerControl1.Panel1.Controls.Add(this.btnNext);
+            this.splitContainerControl1.Panel1.Controls.Add(this.lblMachineName);
+            this.splitContainerControl1.Panel1.Controls.Add(this.txtbMachineName);
+            this.splitContainerControl1.Panel1.Controls.Add(this.btnPrev);
+            this.splitContainerControl1.Panel1.Controls.Add(this.labelControl2);
+            this.splitContainerControl1.Panel1.Controls.Add(this.labelControl3);
+            this.splitContainerControl1.Panel1.Controls.Add(this.lblUser);
+            this.splitContainerControl1.Panel1.Controls.Add(this.txtID);
+            this.splitContainerControl1.Panel1.Controls.Add(this.lblLineNumber);
+            this.splitContainerControl1.Panel1.Controls.Add(this.txtbDateValue);
+            this.splitContainerControl1.Panel1.Controls.Add(this.lblSourceCodeFileName);
+            this.splitContainerControl1.Panel1.Controls.Add(this.txtbDataSource);
+            this.splitContainerControl1.Panel1.Controls.Add(this.lblMethod);
+            this.splitContainerControl1.Panel1.Controls.Add(this.txtSourceValue);
+            this.splitContainerControl1.Panel1.Controls.Add(this.txtbUser);
+            this.splitContainerControl1.Panel1.Controls.Add(this.labelControl1);
+            this.splitContainerControl1.Panel1.Controls.Add(this.lblLevel);
+            this.splitContainerControl1.Panel1.Controls.Add(this.txtbLevelValue);
+            this.splitContainerControl1.Panel1.Controls.Add(this.txtbLineNumber);
+            this.splitContainerControl1.Panel1.Controls.Add(this.txtbProcessModuleName);
+            this.splitContainerControl1.Panel1.Controls.Add(this.txtbFileName);
+            this.splitContainerControl1.Panel1.Controls.Add(this.lblSource);
+            this.splitContainerControl1.Panel1.Controls.Add(this.lblProcess);
+            this.splitContainerControl1.Panel1.Controls.Add(this.txtbMethod);
+            this.splitContainerControl1.Panel1.Text = "Panel1";
+            // 
+            // splitContainerControl1.Panel2
+            // 
+            this.splitContainerControl1.Panel2.AutoScroll = true;
+            this.splitContainerControl1.Panel2.Controls.Add(this.splitContainerControl2);
+            this.splitContainerControl1.Panel2.Text = "Panel2";
+            this.splitContainerControl1.Size = new System.Drawing.Size(975, 604);
+            this.splitContainerControl1.SplitterPosition = 386;
+            this.splitContainerControl1.TabIndex = 3;
+            // 
+            // splitContainerControl2
+            // 
+            this.splitContainerControl2.Dock = System.Windows.Forms.DockStyle.Fill;
+            this.splitContainerControl2.FixedPanel = DevExpress.XtraEditors.SplitFixedPanel.Panel2;
+            this.splitContainerControl2.Horizontal = false;
+            this.splitContainerControl2.Location = new System.Drawing.Point(0, 0);
+            this.splitContainerControl2.Name = "splitContainerControl2";
+            // 
+            // splitContainerControl2.Panel1
+            // 
+            this.splitContainerControl2.Panel1.Text = "Panel1";
+            // 
+            // splitContainerControl2.Panel2
+            // 
+            this.splitContainerControl2.Panel2.Controls.Add(this.meSelected);
+            this.splitContainerControl2.Panel2.Text = "Panel2";
+            this.splitContainerControl2.Size = new System.Drawing.Size(386, 604);
+            this.splitContainerControl2.TabIndex = 0;
+            // 
+            // meSelected
+            // 
+            this.meSelected.Dock = System.Windows.Forms.DockStyle.Fill;
+            this.meSelected.EditValue = "You can double click the selected node to copy the text";
+            this.meSelected.Location = new System.Drawing.Point(0, 0);
+            this.meSelected.Name = "meSelected";
+            this.meSelected.Size = new System.Drawing.Size(386, 100);
+            this.meSelected.TabIndex = 2;
             // 
             // MessageDetailsUC
             // 
@@ -443,40 +568,11 @@
             this.Appearance.Options.UseFont = true;
             this.AutoScaleDimensions = new System.Drawing.SizeF(11F, 22F);
             this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
-            this.Controls.Add(this.xtraTabControlMessageInfo);
-            this.Controls.Add(this.lblThreadId);
-            this.Controls.Add(this.txtbThreadId);
-            this.Controls.Add(this.labelControl5);
-            this.Controls.Add(this.txtbProcessId);
-            this.Controls.Add(this.lblMachineName);
-            this.Controls.Add(this.txtbMachineName);
-            this.Controls.Add(this.txtbIndex);
-            this.Controls.Add(this.txtID);
-            this.Controls.Add(this.labelControl4);
-            this.Controls.Add(this.labelControl3);
-            this.Controls.Add(this.labelControl2);
-            this.Controls.Add(this.lblUser);
-            this.Controls.Add(this.lblLineNumber);
-            this.Controls.Add(this.lblSourceCodeFileName);
-            this.Controls.Add(this.lblMethod);
-            this.Controls.Add(this.txtbUser);
-            this.Controls.Add(this.lblLevel);
-            this.Controls.Add(this.txtbLineNumber);
-            this.Controls.Add(this.txtbFileName);
-            this.Controls.Add(this.lblProcess);
-            this.Controls.Add(this.txtbMethod);
-            this.Controls.Add(this.lblSource);
-            this.Controls.Add(this.txtbProcessModuleName);
-            this.Controls.Add(this.txtbLevelValue);
-            this.Controls.Add(this.labelControl1);
-            this.Controls.Add(this.txtSourceValue);
-            this.Controls.Add(this.txtbDataSource);
-            this.Controls.Add(this.txtbDateValue);
-            this.Controls.Add(this.btnPrev);
-            this.Controls.Add(this.btnNext);
+            this.AutoScroll = true;
+            this.Controls.Add(this.splitContainerControl1);
             this.Margin = new System.Windows.Forms.Padding(4);
             this.Name = "MessageDetailsUC";
-            this.Size = new System.Drawing.Size(875, 593);
+            this.Size = new System.Drawing.Size(975, 604);
             this.Load += new System.EventHandler(this.UCMessageDetails_Load);
             ((System.ComponentModel.ISupportInitialize)(this.txtbDateValue.Properties)).EndInit();
             ((System.ComponentModel.ISupportInitialize)(this.txtbDataSource.Properties)).EndInit();
@@ -496,8 +592,21 @@
             ((System.ComponentModel.ISupportInitialize)(this.xtraTabControlMessageInfo)).EndInit();
             this.xtraTabControlMessageInfo.ResumeLayout(false);
             this.xtraTabPageText.ResumeLayout(false);
+            this.xtraTabPageRenderedText.ResumeLayout(false);
             this.xtraTabPageAdditionalInformation.ResumeLayout(false);
             ((System.ComponentModel.ISupportInitialize)(this.memoAdditionalInformation.Properties)).EndInit();
+            ((System.ComponentModel.ISupportInitialize)(this.splitContainerControl1.Panel1)).EndInit();
+            this.splitContainerControl1.Panel1.ResumeLayout(false);
+            ((System.ComponentModel.ISupportInitialize)(this.splitContainerControl1.Panel2)).EndInit();
+            this.splitContainerControl1.Panel2.ResumeLayout(false);
+            ((System.ComponentModel.ISupportInitialize)(this.splitContainerControl1)).EndInit();
+            this.splitContainerControl1.ResumeLayout(false);
+            ((System.ComponentModel.ISupportInitialize)(this.splitContainerControl2.Panel1)).EndInit();
+            ((System.ComponentModel.ISupportInitialize)(this.splitContainerControl2.Panel2)).EndInit();
+            this.splitContainerControl2.Panel2.ResumeLayout(false);
+            ((System.ComponentModel.ISupportInitialize)(this.splitContainerControl2)).EndInit();
+            this.splitContainerControl2.ResumeLayout(false);
+            ((System.ComponentModel.ISupportInitialize)(this.meSelected.Properties)).EndInit();
             this.ResumeLayout(false);
 
         }
@@ -539,5 +648,10 @@
         private DevExpress.XtraTab.XtraTabPage xtraTabPageText;
         private DevExpress.XtraTab.XtraTabPage xtraTabPageAdditionalInformation;
         private DevExpress.XtraEditors.MemoEdit memoAdditionalInformation;
+        private DevExpress.XtraTab.XtraTabPage xtraTabPageRenderedText;
+        private DevExpress.XtraRichEdit.RichEditControl recMessageDetails;
+        private DevExpress.XtraEditors.SplitContainerControl splitContainerControl1;
+        private DevExpress.XtraEditors.SplitContainerControl splitContainerControl2;
+        private DevExpress.XtraEditors.MemoEdit meSelected;
     }
 }
